Card-ify login components
