+++ src/components/landing/pages/login/login.scss
@@ -1,3 +1,3 @@
 .social-button-container {
-  min-width: 30%;
+  width: 100%;
 }
